Extract MappedBuffer class into RF2SharedMemoryReader.cs, similar to Crew Chief.  Makes it easier to compare. Make meaningless start on sending shared data!
