+++ Monitor/rF2SMMonitor/VC16/rF2SMMonitor.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30204.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "rF2SMMonitor", "rF2SMMonitor.csproj", "{17FC2626-EA21-43ED-BC84-BE0B8545FC2D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{17FC2626-EA21-43ED-BC84-BE0B8545FC2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{17FC2626-EA21-43ED-BC84-BE0B8545FC2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{17FC2626-EA21-43ED-BC84-BE0B8545FC2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{17FC2626-EA21-43ED-BC84-BE0B8545FC2D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {70C8E4C0-3F5C-41B8-9DEE-01267B1C398B}
+	EndGlobalSection
+EndGlobal
